The big package split-up!
This commit introduces ten new "sourceforge-*" subpackages (in
addition to the traditional "sourceforge" one).  I performed a minor
amount of testing on it.  The packages build, they even install, but I
trust there are lots of areas where they do not work as they should
(or just plain do not work).

Yay!


git-svn-id: svn+ssh://scm.fusionforge.org/srv/svn/fusionforge/trunk@681 9d84d37e-dcb1-4aad-b103-6f3d92f53bf6