First commit on the Branch_plugins branch.
Currently implemented: a table containing a list of installed plugins,
a table for a list of "this group uses that plugin" relations, one for
"this user uses that plugin" relations, and a usesPlugin() method for
the User and Group classes (in PHP).  Still to be done: a sample
plugin, a page listing all installed plugins, scripts to initialise
plugin_specific tables.


git-svn-id: svn+ssh://scm.fusionforge.org/srv/svn/fusionforge/trunk@985 9d84d37e-dcb1-4aad-b103-6f3d92f53bf6