Large commit bringing the Mediawiki plugin into shape.
Renamed the User class to GFUser, in order to prevent a name clash with
Mediawiki.  Also renamed $conn to $gfconn, for similar reasons.

The Mediawiki plugin adds an optional tab to projects that enable it.
Under that tab is an iframe containing Mediawiki with a light theme.
Authentication is propagated from GForge to Mediawiki.


git-svn-id: svn+ssh://scm.fusionforge.org/srv/svn/fusionforge/trunk@6348 9d84d37e-dcb1-4aad-b103-6f3d92f53bf6